+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Chg: Internally, `archived` property was renamed to `in_trash`.
 - Chg: Renamed `.content` to `.children` of a page for more consitency.
 - Chg: Use Property again to consistently refer to the columns of a database.
+- Chg: Make use of [pendulum](https://pendulum.eustace.io/) to represent `DateRange` values.
 
 ## Version 0.4, 2024-02-14